// src/data/interface.tsx
export interface CarouselInterface {
  id: string;
  heading: string;
  para: string;
  image: string;
  buttonText: string;
}

export interface ServicesInterface {
  id: string;
  image: string;
  heading: string;
  para: string;
  color: string;
  img1: string;
  backGround: string;
}

export interface GalleryInterface {
  image: string;
}

export interface ExpertTeamInterface {
  id: string;
  image: string;
  heading: string;
  paragraph: string;
}

export interface TestimonialInterface {
  id: string;
  image: string;
  name: string;
  head: string;
}

export interface OurBlogsInterface {
  image: string;
  month: string;
  monthName: string;
  title: string;
  paragraph: string;
}